:art: move Providers into react-ui folder & more
