+++ client2/src/README.md
@@ -1,0 +1,3 @@
+# Front End
+
+<img src="https://raw.githubusercontent.com/do-kevin/Zephyr-Node-docs/main/assets/zephyr-node-0.2.svg" />
