+++ client2/src/core/config.ts
@@ -1,4 +1,5 @@
 import { injectable } from "inversify";
+import "reflect-metadata";
 
 @injectable()
 export class Config {
